* moves addFieldTypes to new file * creates new dateAndTime field
